Overnight facet program: is mindedness ONE axis or several? (prereg + 4 stages)
Motivated by the decode: pooled v_mind reads as ethics/faith, so the 6 pooled attributes may be
separable sub-concepts. Frozen facets: pain/emotion/consciousness/soul/cognition/agency, each vs
the same physical baseline. S1 screen (1 template) -> S2 wide (3 templates + polarity-orth +
split-half + floor, ALL facets, no auto-selection to avoid researcher dof) -> S3 decode each facet
direction -> S4 steer consciousness vector with all 6 facets as DVs at low alpha, 5 random seeds.
Predictions incl. H-soul-special (different entity gradient, religious vocabulary) and
H-self-localised (self-outlier largest on consciousness facet).
Also fixes the Qwen3.5 steering dtype bug (cast steer vector to tensor dtype; bf16 hybrid conv).

+++ mvp/mindedness_steer.py
@@ -102,11 +102,13 @@ def main():
         def __init__(self, vec, alpha):
             self.h = None
             if alpha != 0.0:
-                v = torch.tensor(vec, dtype=torch.float16, device=DEVICE)
+                v = torch.tensor(vec, dtype=torch.float32, device=DEVICE)
                 def hook(m, i, o):
                     t = o[0] if isinstance(o, tuple) else o
+                    # cast to the TENSOR's own dtype: Qwen3.5 keeps bf16 in its hybrid conv
+                    # layers, and bf16 + fp16 promotes to fp32 -> conv1d dtype mismatch.
                     scale = t.norm(dim=-1, keepdim=True) * alpha
-                    t = t + scale * v
+                    t = t + (scale * v.to(t.dtype))
                     return (t,) + o[1:] if isinstance(o, tuple) else t
                 self.h = model.layers[SL].register_forward_hook(hook)
         def __enter__(self): return self
